+++ eth_defi/utils.py
@@ -174,3 +174,4 @@ def get_url_domain(url: str) -> str:
         return parsed.hostname
     else:
         return f"{parsed.hostname}:{parsed.port}"
+